refactor(state): extract TTS announcer subsystem into mixin (#1271)
First-increment split of the game/state.py God-Object (#1271). Moves the
cohesive TTS / spoken-announcement cluster (configure_tts, disable_tts, the
announce_* family, _lang, _tts_announce, _announce_reveal — Issues #447 /
#471 / #840 / #841 / #842 / #1211) out of GameState into a new
TtsAnnouncerMixin in game/state_tts.py.

Behavior-preserving: GameState now inherits the mixin so its public API and
every caller/test are unchanged. The _tts_* state is initialized via the
mixin's _init_tts_state(), called from GameState.__init__.

state.py: 2913 -> 2512 lines (-401).
